Refactor: handle_hidden으로 숨김관련 함수 통합
