+++ lib/pages/main_page.dart
@@ -361,6 +361,7 @@ class _MainPageState extends State<MainPage> with WidgetsBindingObserver {
 
   @override
   Widget build(BuildContext context) {
+    debugPrint("build invoked!!");
     return Scaffold(
       appBar: AppBar(
         elevation: 0,
