+++ lib/widgets/post_preview.dart
@@ -7,6 +7,7 @@ import 'package:new_ara_app/translations/locale_keys.g.dart';
 import 'package:new_ara_app/models/article_list_action_model.dart';
 import 'package:new_ara_app/utils/time_utils.dart';
 import 'package:new_ara_app/providers/blocked_provider.dart';
+import 'package:new_ara_app/utils/handle_hidden.dart';
 import 'package:provider/provider.dart';
 
 class PostPreview extends StatefulWidget {
@@ -52,7 +53,7 @@ class _PostPreviewState extends State<PostPreview> {
                               .contains(widget.model.created_by.id))
                       ? LocaleKeys.postPreview_blockedUsersPost.tr()
                       : getTitle(widget.model.title, widget.model.is_hidden,
-                          widget.model.why_hidden, context.locale),
+                          widget.model.why_hidden),
                   style: TextStyle(
                     // TODO: 아래 코드는 iOS 심사 통과를 위한 임시 방편. 익명 차단이 BE에서 구현되면 제거해야함 (2023.02.29)
                     color: (widget.model.is_hidden ||
@@ -232,45 +233,5 @@ class _PostPreviewState extends State<PostPreview> {
   // TODO: 아래 코드는 iOS 심사 통과를 위한 임시 방편. 익명 차단이 BE에서 구현되면 제거해야함 (2023.02.29)
   bool isAnonymousIOS(ArticleListActionModel model) {
     return (Platform.isIOS && model.name_type == 2);
-  }
-
-  /// 게시글 정보를 입력받고 그에 상태에 따라 적절한 제목을 리턴하는 함수.
-  /// UserViewPage, UserPage, PostListShowPage, PostViewPage에서 사용함.
-  String getTitle(String? orignialTitle, bool isHidden, List<dynamic> whyHidden,
-      Locale locale) {
-    // 숨겨진 글이 아닌 경우
-    if (isHidden == false) {
-      return orignialTitle.toString();
-    }
-    // 숨겨졌으나 why_hidden이 지정되지 않은 경우. 혹시 모를 에러 방지를 위해 추가함.
-    else if (whyHidden.isEmpty) {
-      return LocaleKeys.postPreview_hiddenPost.tr();
-    }
-
-    // TODO: 새로운 사유가 있을 경우 코드에 반영하기.
-    late String title;
-    switch (whyHidden[0]) {
-      case "REPORTED_CONTENT":
-        title = LocaleKeys.postPreview_reportedPost.tr();
-        break;
-      case "BLOCKED_USER_CONTENT":
-        title = LocaleKeys.postPreview_blockedUsersPost.tr();
-        break;
-      case "ADULT_CONTENT":
-        title = LocaleKeys.postPreview_adultPost.tr();
-        break;
-      case "SOCIAL_CONTENT":
-        title = LocaleKeys.postPreview_socialPost.tr();
-        break;
-      case "ACCESS_DENIED_CONTENT":
-        title = LocaleKeys.postPreview_accessDeniedPost.tr();
-        break;
-      // 새로운 whyHidden에 대해서는 숨겨진 게시글로 표기. 이후 앱에서 반영해줘야 함.
-      default:
-        debugPrint("\n***********************\nANOTHER HIDDEN REASON FOUND: ${whyHidden[0]}\n***********************\n");
-        title = LocaleKeys.postPreview_hiddenPost.tr();
-    }
-
-    return title;
   }
 }
